+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * [UCX-3898] Can't unmute user from User details flex panel
 * [UCX-3913] Show SideNav More Unreads immediately instead of after scroll
+* [UCX-3912] Clear the message box when send button clicked
 
 
 ## 1.0.0.beta12 (2018-04-09)